build table view to be transferred
