+++ src/main/java/com/aneesh/builder/TableBuilder.java
@@ -1,8 +1,16 @@
 package com.aneesh.builder;
 
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileReader;
+import java.io.IOException;
+
 public class TableBuilder {
 
     public String[] columnNames;
+    public static final String NO_FILE_MESSAGE= "No files to display";
+    public String [][] tableData;
+
 
 
     public String[] defineColumnNames() {
@@ -16,5 +24,41 @@ public class TableBuilder {
         columnNames[6] = "BIC (57a)";
         return columnNames;
     }
+
+    public void buildData(File[] mt940Dir) throws IOException {
+
+        String[][] fileContent;
+
+        if(mt940Dir.length == 0){
+            fileContent = new String[1][1];
+            fileContent[0][0] = NO_FILE_MESSAGE;
+        }
+        else{
+            fileContent = new String[mt940Dir.length][7];
+
+            int i = 0;
+            for(File f: mt940Dir){
+                BufferedReader reader = new BufferedReader(new FileReader(f.getAbsolutePath()));
+                String line = reader.readLine();
+                int j = 0;
+
+                fileContent[i][j] = f.getName();
+                j++;
+                while(line != null){
+
+                    fileContent[i][j] = line;
+                    j++;
+                    line = reader.readLine();
+                }
+                i++;
+            }
+
+        }
+
+        tableData = fileContent;
+
+
+    }
+
 
 }
