+++ src/main/java/com/aneesh/gui/SwiftGui.java
@@ -4,7 +4,6 @@ import com.aneesh.builder.TableBuilder;
 
 import javax.swing.*;
 import java.awt.*;
-import java.io.*;
 
 public class SwiftGui extends JFrame {
 
@@ -14,9 +13,7 @@ public class SwiftGui extends JFrame {
     private static JTable table;
     private static int guiWidth = 600;
     private static int guiHeight = 400;
-    public static final String NO_FILE_MESSAGE= "No files to display";
-    String [][] tableData;
-    private static TableBuilder tableBuilder;
+    public static TableBuilder tableBuilder;
     private String[] columnNames;
 
     private SwiftGui(){
@@ -30,42 +27,6 @@ public class SwiftGui extends JFrame {
 
     }
 
-    public void buildData(File[] mt940Dir) throws IOException {
-
-        String[][] fileContent;
-
-        if(mt940Dir.length == 0){
-            fileContent = new String[1][1];
-            fileContent[0][0] = NO_FILE_MESSAGE;
-        }
-        else{
-            fileContent = new String[mt940Dir.length][7];
-
-            int i = 0;
-            for(File f: mt940Dir){
-                BufferedReader reader = new BufferedReader(new FileReader(f.getAbsolutePath()));
-                String line = reader.readLine();
-                int j = 0;
-
-                fileContent[i][j] = f.getName();
-                j++;
-                while(line != null){
-
-                    fileContent[i][j] = line;
-                    j++;
-                    line = reader.readLine();
-                }
-                i++;
-            }
-
-        }
-
-        tableData = fileContent;
-
-
-    }
-
-
 
     public static synchronized SwiftGui getGuiSingleton(){
         if(guiSingleton == null){
@@ -76,7 +37,7 @@ public class SwiftGui extends JFrame {
         return guiSingleton;
     }
 
-    private void setDefaultGuiProperties() {
+    private void setProperties() {
         guiSingleton.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         guiSingleton.setSize(guiWidth,guiHeight);
         guiSingleton.setVisible(true);
@@ -88,20 +49,16 @@ public class SwiftGui extends JFrame {
     public void buildTableView() {
 
 
-        table = new JTable(tableData, columnNames);
+        table = new JTable(tableBuilder.tableData, columnNames);
         table.setAutoCreateRowSorter(true);
         table.setPreferredScrollableViewportSize(new Dimension(500  ,300));
         table.setFillsViewportHeight(true);
         JScrollPane jScrollPane = new JScrollPane(table);
         add(jScrollPane);
 
-        setDefaultGuiProperties();
+        setProperties();
 
     }
 
-    public static JTable getTable() {
-
-        return table;
-    }
 }
